# let fpgrowth and fpmax work directly with sparse input