@pep8speaks 2019-11-06
Hello @dbarbier! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:





* In the file [`mlxtend/frequent_patterns/tests/test_fpbase.py`](https://github.com/rasbt/mlxtend/blob/7948c1fd71b028cd1089c1fd7ad75d9c26924f90/mlxtend/frequent_patterns/tests/test_fpbase.py):

> [Line 189:80](https://github.com/rasbt/mlxtend/blob/7948c1fd71b028cd1089c1fd7ad75d9c26924f90/mlxtend/frequent_patterns/tests/test_fpbase.py#L189): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (91 > 79 characters)