@rasbt 2019-11-06
Thanks so much again for another great PR. The improvements are quite substantial with `sparse=True`. Many thanks!